requirements.txt
python-dotenv==1.0.0
libmagic==1.0
langchain==0.0.284
python-magic==0.4.27
unstructured==0.9.2
python-magic-bin==0.4.14
tiktoken==0.4.0
OpenAI == 0.28.0
faiss-cpu==1.7.4
streamlit==1.22.0